+++ routers/web/admin/auths.go
@@ -10,7 +10,6 @@ import (
 	"net/http"
 	"regexp"
 
-	"code.gitea.io/gitea/models"
 	"code.gitea.io/gitea/models/login"
 	"code.gitea.io/gitea/modules/auth/pam"
 	"code.gitea.io/gitea/modules/base"
@@ -386,7 +385,7 @@ func EditAuthSourcePost(ctx *context.Context) {
 	source.IsSyncEnabled = form.IsSyncEnabled
 	source.Cfg = config
 	if err := login.UpdateSource(source); err != nil {
-		if models.IsErrOpenIDConnectInitialize(err) {
+		if oauth2.IsErrOpenIDConnectInitialize(err) {
 			ctx.Flash.Error(err.Error(), true)
 			ctx.HTML(http.StatusOK, tplAuthEdit)
 		} else {
